Derive a parameter-free version of the generic recursion principle.
I can't figure out how to properly lift the parameterised version, and I can re-derive it from a
normal recursion principle, so for the moment, it does rather look as if the work to
derive the parameterised principle for generic terms was unnecessary.

Being able to the do the re-derivation is a step forward though, so that's a little
consolation.