Terms and labelled terms now derived from generic terms.
This is a big change that points to the implementation of a complete
nominal datatype package.  The use of generic terms means that we only
need to do one hairy quotient, and one hairy derivation of a recursion
principle.

The recursion principle in question features a new "FCB" (one of the
side-conditions that need to be proved in order to get a recursive
function admitted).  Though I think it's better, it can be a little
fiddlier to prove.  It's also stronger, so in theory, there may be
reasonable functions that the new FCB rejects. It should certainly be
possible to automate the bulk of the new fiddly proofs.