Fix parsing in TotalDefn so that it doesn't depend on global context.
The DIV_LESS_I lemma that it proves was breaking when the parser had been primed
with the syntax of categoryTheory.  The syntax in the latter might be a mistake
but it shouldn't be able to subvert core functionality.  This bug couldn't be
seen under Poly/ML because TotalDefn is pre-loaded into the HOL image before
functorScript.sml runs, and the DIV_LESS_I lemma was thus already proved.