A general tool to build HOL heaps.
This tool reads standard input, expecting first a line specifying the name
of the new heap, and then a list of dependencies (which will be loaded
using the SML load function.

These arguments can't be on the command-line because HOL heaps use the
command-line to specify SML files that are to be used.  Indeed, the
expectation is that this genbuild.ML will be just such a file.

This file only builds the object file; there needs to be a subsequent
call to the C compiler to link this, creating the final executable.