Illustration of new local heap functionality.
This is still not a very polished workflow.  In this example
(computability/lambda) it turns a 3 minute build into a 2 minute
build, so there's a definite advantage to it, even on what is not a
huge example.

The construction should always be the same so I may just move it into
Holmake.  Because it's not yet polished, I'll hold off documenting it
for the moment.

If you run the resulting heaps as they are, you don't have the
pretty-printing loaded.  Thus you will want to do something like

  ./myheap <holdir>/tools-poly/prelude.ML

in order to get a proper interactive session.