Make the hol executable use a directory's HOLHEAP, if given.
This is achieved with a heapname executable that runs before hol
starts, returning the name of the appropriate heap for the current
directory.  This determination is made by looking at the local
Holmakefile, if there isn't one, or if it doesn't specify a HOLHEAP
variable, then the usual bin/hol.builder heap is used.

This also works for the hol.noquote executable (though it's hard to
believe anyone uses this).