Shell scripts for assessing the state of Docfile documentation
I'm not sure how accurate these are; I'm not an expert scripter.
Patches welcome!

It seems like everything document is actually implemented somewhere, but
there are about 18000 things undocumented.

It might be worthwhile starting a list of available functionality that
we do not want to document to filter that number down some.

Currently I'm not checking to make sure types in the documentation match
the code, or that the SEEALSO links aren't broken. But these would be
nice functionality to have.