An operational semantics for a (practical?) ML subset.
A small step operational semantics and type system for a small, purely
functional subset of ML (more in the style if OCaml than SML).
Conspicuously absent is let-polymorphism in the type system, and most
syntactic sugar.  In an effort to be practical, it supports mutually
recursive functions and datatypes, and nested pattern matching.

Preservation and progress have been proved for expression evaluation,
but not for declaration evaluation.

MiniMLScript.sml is generated by Lem
(http://www.cl.cam.ac.uk/~so294/lem) from miniML.lem (although the
current release version (0.2) won't do this right) which should be
considered the definitive source.  MiniMLScript.sml should not be
hand-edited.