start a loggable kernel whose theorems have proofs attached
originally these commit messages:
--
start work on a liberally logging kernel

The idea is to log HOL4's primitive (including derivable) inferences as
they happen, and then deal with converting this log to one using only
Opentheory primitives later. One advantage might be that it should
require less extensive changes to the HOL4 kernel.

This commit is a minor preparatory step: split the OpenTheoryMap into
two parts, one that will happen prekernel (so that Thm can see the
Opentheory names to log) and the other that happens after Theory (so it
can use LoadableThyData). Of course, Thm should only ever look at the
map well after various later theories have populated it (with
side-effects) using the second interface.
--
Actually, better to keep logging after the kernel

Which is possible if we store logging information in the form of
"proofs" attached to the theorems. The only visible change is that the
proof type and functions like get_proof, delete_proof etc. will be
visible in Thm, but that appears to be fine (so it's a superset of
FinalThm rather than matching exactly). This is the way Joe logs
HOL-Light proofs - I intend to emulate fairly closely.

The main reason to do it this way is that not every theorem produced
needs to be logged. Indeed logging facility will probably have a log_thm
: thm -> unit, which could be linked to save_thm (at the expense of
touching Theory).

Thm doesn't need to know whether logging is happening or not. But it may
pay to let it know (e.g. via a flag in Globals) later if storing proofs
all the time is too memory-intensive.