start work on log_thm
Contrary to the initial plan of having an intermediate extended article
file format, it seems feasible to do the compilation to the OpenTheory
virtual machine on the fly. Being outside of the kernel (and thus in the
scope of boolTheory, Drule, etc.) makes this possible. However, we must
be careful to avoid implementing an inference rule in terms of itself
(and thus sending log_thm into a loop).

SUBST is the first rule to pose a challenge (and probably won't be
the last), so I've left it for now. Next step is probably to get rid of
TODO_prf (by adding proof cases for the remaining rules in Thm).