implement prim_type_definition for opentheory
Currently uses backward proof to generate the theorem the virtual
machine should produce. Could probably be replaced with a proforma
theorem (as in many other commands) if that would be beneficial.

We only make definitions (of types or constants) on the virtual machine
if the theorems generated by prim_type_definition or
prim_constant_definition get used in the logging of an explicitly
exported theorem. This might be good enough, but it might also be nice
to allow a user to explicitly log a definition without actually using
those theorems (maybe the user doesn't know when they're getting used).
(Easy to implement: just log the theorem without calling "thm" (i.e.
without saving it as an article export) at the end.)