Simplify WF_uR proof, hoping to make Moscow ML builds accept it.
Given that the Moscow ML failures occur non-deterministically, because
METIS_TAC is itself non-deterministic in behaviour, it's possible this
isn't really fixed.  Nonetheless, on my Mac, I could repeatedly get it
to fail before "fixing" it.  I then got it to build successfully with
the given change (which just makes the relevant terms smaller through
the use of abbreviations).