major reworking of treatment of definitions
Since an article needs to define things before it can use them, it is
nice to be able to log definitions in the order they happen in the
kernel. So I have added a generic definition_callback to Thm which is
called when definitions happen, and Logging installs a function that
collects definitional theorems. They can then be logged in order by a
call to log_definitions() right after the definition is made in the
script file being logged.