rewrite SUBST_prf to unconditionally rename bound variables
if they occur free anywhere in the substitution or the source theorem.
if they don't occur free, then hopefully prim_variant and subst will
notice the opportunity to do nothing.

This is prompted by the situation where the map contains theorems
with free variables in its hypotheses that are bound in the template.
Our OpenTheory implementation of SUBST now differs from the HOL4
implementation in that it doesn't return the template exactly, but
something alpha-equivalent. But OpenTheory theorem objects are supposed
to be equal up to alpha-equivalance...