Fix cl example for changes to Hol_datatype.
Also strip the "old" tags from constants if necessary when defining them
by specification. (The constant might already have been deleted by the
time the virtual machine is going to define it in terms of Hilbert
choice, but we need the constant and not just its name to construct the
alternative definition c = @x. <foo> so it matters whether the constant is
uptodate.)

Minor tweak to cl.thy; will now try uploading it to the gilith repo.