make foo.art export the same theorems as fooTheory.sml
instead of exporting after each call to store_thm (which misses theorems
created as definitions or axioms), we map export_thm over the current
theorems, defs, etc. when export_theory is called.

this is good default behaviour, but there should be an option for more
fine-grained control over the exports of an article produced from a
script.

also, remove the mistaken import of base in the cl.thy file.