Fixes an error in METIS_TAC for boundary cases of genvars.
The arity function in mlibTermorder.sml made use of String.compare
to compare two strings, including the names of genvars.

The problem with this is that while

> String.compare ("%%genvar%%999","%%genvar%%998");
val it = GREATER : order

sometimes we have

> String.compare ("%%genvar%%1000","%%genvar%%999");
val it = LESS : order

This is important, and has led to METIS_TAC failing when it would have
worked if the genvar numbers had been one less or one more.

Joe Hurd, the author of METIS_TAC, has been informed and has agreed
with this change:

"Thank you for such a detailed bug report together with a thoughtful
analysis of the underlying problem and a suggested patch. I completely
agree with your reasoning, and I think your patch will help to make
METIS_TAC more deterministic (an important consideration for an
interactive theorem prover).

Please commit your patch with my blessing."