hooks for logging names without global mappings
By default (i.e., after calling loggingBossLib.new_theory) hooks are set
up to translate a constant or type called thy$name, as long as thy is
the current theory, to the opentheory name ([thy],name).

Thus when logging a new theory, one does not have to set up any mappings
in the OpenTheoryMap explicitly. They can be put there, though, of
course, and the handler could be set to update the map as well as
returning a default name.

Using the new defaults, the names in the combinatory logic example are
all now in the "cl" namespace. To use different namespaces or to rename
certain constants (and to use "combinatoryLogicExample", which is not
the HOL4 theory name) would require a companion file specifying this
information.