define arithmetic$ZERO and redefine num$0 opentheory mapping
The main issue is that we want both of these constants to point to
Number.Numeral.zero, but the OpenTheoryMap assumes one-to-one mapping
situations. Another issue is that we want Number.Numeral.zero to point
back to num$0 but arithmetic$ZERO is defined later (and the map retains
the last insertion.)

Solution here is to insert num$0 again after arithmetic$ZERO.
Downside is that a warning is printed saying that a duplicate mapping is
being inserted (probably every time arithmeticTheory is loaded).

Upside is that clScript.sml is now completely unchanged (since before
experimentation with opentheory logging) but is still able to be logged
to an OpenTheory article, with reasonable defaults for all necessary
choices.

The warning could be removed by suppressing the warning when the new
mapping is to the same string, by special-casing zero somewhere
(probably OpenTheoryMap), or by implementing a real many-to-many (or
one-to-many) map situation (hardest).