expose hook for translating names when reading an article
So it is not necessary (although maybe often recommended) to update the
OpenTheoryMap before reading an article. Rather, one passes a function
that computes the desired HOL4 name from an OpenTheory name in the
article.

Currently, the global map is not considered by default first, but we
provide functions that look it up that can be used as these hooks when
defining a reader that requires no special treatment.

Updated the OTcl example to not touch the global map.
There is still a bit of special-casing for the @ind_type constants
(because the names are not valid ML identifiers) but that will hopefully
go away when those constants are actually deleted in the appropriate
places...