delete unused constants before exporting OTclTheory
Now [constants "OTcl"] returns the same list as [constants "cl"].
Also, no more need to specially test for the constants whose names
contain "ind_type".

But the definitions made by OTcl still mostly reappear (in a nicer form)
in its lone theorem "exports" (a conjunction of the article exports).
So it's not clear whether it should be storing the definitions as well.