bugfix for logging MP proofs
The first theorem to MP is allowed to be a negation. In that case we now
explicitly rewrite it to an implication before doing the inferences that
simulate MP on implications. Any better ways to do that rewrite than
SUBS_OCCS NOT_DEF followed by BETA_CONV?

Also, we now construct the desired theorem in HOL (since we have a
deductAntisym rule in the logging kernel) before logging it, rather than
only doing the construction on the virtual machine: thus, more
hash-consing of theorems in the dictionary.