unfortunate bugfix for logging SUBST proofs
The hypotheses of the theorem returned by SUBST includes the hypotheses
of all the theorems in the [var|->equality theorem] substitution, even
if some of those theorems are never used (e.g. because the var doesn't
appear in the template). Therefore in the implementation of SUBST for
the OpenTheory kernel, we now explicitly add any hypotheses to the
resulting theorem (using ADD_ASSUM) that were not actually needed in its
proof.

Any suggestions on adding these hypotheses more efficiently would be
welcome. Should SUBST even be adding them (in HOL4)? (Documentation says
yes, but was this discussed?)

Also, we now prove the theorem in HOL4 first then log it, rather than
writing virtual machine commands directly. Excuses for doing this
include: easier to debug (compare the theorem you do prove against the
one you wanted to), and makes more use of hash-consing (all sub-theorems
will go through the dictionary via recursive calls to log_thm).