bugfix for logging SUBST proofs with variable reuse
When substituting at positions marked by a variable that happens to
appear free in the conclusion of the source theorem, the conclusion may
be identitical to the template. In this case, the previous code assumed
no substitution was necessary since the source already matched the
template. The new code explicitly introduces fresh variables to mark
positions, updating the map and template accordingly, before doing the
actual substitution. Thus variables marking substitution positions will
differ from the source theorem and therefore be found. (Performance
suggestions welcome, as always.)