Make Systeml.sig the same in Moscow ML and Poly builds.
This additional uniformity allows me to delete the
tools-poly/Holmake/Systeml.sig file.  This then requires a bunch of
changes to compilation of the early tools which have to point at
tools/Holmake/Systeml.sig rather than the version in
tools-poly/Holmake.

This should fix the configuration errors on MoscowML introduced by
0a0988fc.