Forces complete unification of types in term combinations.
Modified the type inference in Preterm.sml to force the complete
unification of the types compared when a term function is applied
to its argument.  Previously, this unification could have been postponed
if the unification involved a higher order match.  But that postponement
meant that after the unification, it may not be true that the type of the
term function was in fact of the form ``:'a -> 'b``.

Also adjusted a couple of error messages in Thm.sml.