Use a "numheap" between hol.builder0 and hol.builder to speed builds.
The numheap heap is stored in num/termination (as per the comment in
build-sequence). Also fix some minor problems with dependency handling
for heaps in Holmake.  Use of the new heap requires Holmakefiles in
the relevant directories to explicitly say they want the heap to be
used.

Holmake also now has to allow Holmakefiles to override command-line
choice of poly.  This command-line option is used by build to move
between "phases".  This is perhaps not ideal, as most command-line
options take precedence over what may appear in Holmakefiles.