Restricted all eta conversions to require rank equality.
This affects primarily Type.eta_conv_ty, but also Term.eta_conv and
Term.ty_eta_conv.

The issue is that in the type \'a. 'a 'M, 'M and 'a have to agree in
terms of the kind that 'M expects, but it is possible that might have
lower rank.  If so, then the eta conversion is not proper,
since 'M alone is a larger function, operating on a larger domain,
than \'a. 'a 'M. Therefore the two are not equal, and should not be
converted into one another.