add a more substantial test for HolQbfLib
I think this is the only thing in the selftest that actually makes use
of the minisat integration. It works (reasonably quickly) with
HolQbfLib.prove, but loops with HolQbfLib.decide because
Canon.PRENEX_CONV loops (so I commented that part of the test out).
Source of test: http://www.qbflib.org/benchmark_detail.php?idBench=1476.