specify path to minisat in holqbf.sml
Unfortunately, a full path is needed, so it's not sufficient to put
minisat in the $PATH variable on the server. So one would have to make a
version of holqbf for each server (and edit the executable string when
building it).

Also, it would be nice if we had functional record updates in SML...

Finally, a note: the version of minisat included with HOL4 does not seem
to be available on the minisat website. The proof-logging version on the
website (with version number 1.14) nevertheless behaves differently
(e.g. accepts different options, prints different help message).