Some more theorems about LIST_REL.
LIST_REL_MAP1 uses composition and allows for anything in the R
position.
LIST_REL_MAP2 requires an abstraction in the R position and so can
directly apply the f to an abstracted variable.  There is a version of
this that looks like

  LIST_REL R l1 (MAP f l2) = LIST_REL (combin$C ($o $o R) f) l1 l2

and this can be made to do the right thing when R is an abstraction
with subsequent rewrites from combinTheory, but it looks so confusing
that I thought we might as well restrict it to the case when R is
already an abstraction.