Add new rewrite about abstractions and function composition.
This one parallels the existing o_ABS_R.  To my surprise, the
simplifier currently preserves bound variable names when this rewrite is
applied, even though the two x's can well be of different types, and
thus different bound variables in a pretty strong sense.

For example,

  SIMP_CONV (srw_ss()) [o_ABS_L] ``(\b. b /\ p) o FST``

gives back

   |- (λb. b ∧ p) o FST = (λb. FST b ∧ p)

The b on the left is a boolean, the b on the right is a pair.  This is
a bit odd, perhaps even a bug.  When the types are the same this is
clearly the right thing, but still.  The bug, if it is one, will be
in the implementation of HO_PART_MATCH.