a standalone tactic of numLib.DECIDE
This currently isn't practical because an OpenTheory term coming in
wouldn't have NUMERAL tags on it, and DECIDE borks in that case.
Probably the solution is to have the article reader automatically figure
out where NUMERAL tags would be necessary and add them on...

This commit also adds a Holmakefile to build the new tactic as well as
the skico tactic (which previously was built manually).