Verified the LazyPairingHeap example.
This has a nested recursion that I failed to prove termination for
directly (as in Konrad's 91 function and lambda-calculus substitution
examples).  Instead I add a check around the nested call

let h3 = nested_call in
  if size h3 < ... then
    outer_call h3 ...
  else
    ARB

Then Define works, and I can prove that the size is always ok, and
rewrite away the check.