In experimental kernel, faster inst_all and fixed bug in free_vars_lr.
The bug in free_vars_lr was the source of a bug reported by Ramana Kumar
as causing a problem with "Define"; many thanks, Ramana!

The new version of inst_all is about twice as fast as before, as it
merges the algorithms for "subst" and "inst_rk_kd_ty" together, for
a single traversal of the term to which the substitution is applied.