new treatment of declaration variables in miniml compiler
the idea was to only keep track of variables that will end up visible at
the top-level and throw the names of the rest away. (was already not
keeping them on the stack.) not sure if that's really happening now, but
the tests still work. a further improvement would make declarations of
variables with the same name as previously declared variables overwrite
those locations on the stack.