More MiniML printer improvements.
Most of the time to build the token list was taken to decide if an SML name
needs the op prefix (e.g., + - mod before, etc), and changing from a linear to a
binary search made a huge difference.  Now 80% of the time is spent in
converting the token list to a string, and the AST to token list is fast.