notes on new approach to miniml compiler correctness
now trying to define a relation between source and intermediate language
expressions that characterises equivalent expressions (and prove a
theorem that related expressions evaluate to the same observable
values). the relation is parameterised by a "global knowledge" of
"equivalent" values for feeding as arguments, inspired by Hur, Dreyer,
Neis, Vafeiadis, but a lot more hacky at the moment. other approaches
are very briefly mentioned in comments.