A benchmarking tool for the bytecode (requested by Ramana)
bytecode_testerLib defines a function run_bytecode, which given
a HOL term with a list of bytecode instructions, e.g.

  ``[Stack (PushInt 4); Stack (PushInt 400); Stack Pop]``

and a number of repetitions to perform, e.g. 100 million, prints
the content of the resulting stack and timing information.
To run the above example 100 million times takes 0.127 seconds
on my laptop.