Started working on a verified lexer generator.
So far I have defined the correct tokenisation of a string, given a lexer
specification.  I have verified that the lexer's run-time driver finds that
tokenisation (iff one exists) when the driver is given a correct DFA for the
regular expressions in the spec.  I haven't started on a verified DFA builder
yet.