slow progress on compiler correctness
mainly trying to prove that adding/removing junk from environments is
allowed by the semantics of the intermediate language. this is
significantly complicated by closures, and relies on the intermediate
language semantics forcing canonical environments for its closures.

defined, programmatically, EVERY_Cexp and EVERY_Cv for mapping over
those types with a predicate. a little complicated because of the mutual
recursion between them.