Adding simple case of subset types to quotient package.
Given a predicate P:'a -> bool for which ?x:'a. P x,
there is a simple and common kind of quotient that follows,
where the partial equivalence relation is

R x y = P x /\ P y /\ (x = y)

Two new entry points are added to the quotient library to ease
creating such quotients.