more slow progress on compiler correctness
Made force_dom take a default value so that we can still know something
good about the values in the environments stored in closures in badly
typed terms.

Tried fixing EVERY_Cexp to be more (mutually) recursive, but termination
proofs become impossible. Instead, made it easy to automatically define
instances of EVERY_Cexp P from the interesting equations for P (i.e.
automatically fill out the recursion).

Fixed semantics of Cevaluate for calling recursive closures: was putting
the index rather than the name as the last component of the CRecClos
generated for the environment in which to evaluate the body. Also need
to canonicalise the environment for those CRecClos's. But now thinking
perhaps it would be better to simply check closures have good
environments everywhere in Cevaluate, and also force good environments
for anything it produces... i.e. move the proof of that fact into the
definition.