Introduce the idea of a "canonical" wellorder.
These correspond to the "ordinals" defined in the original HOL88
wellordering library by John Harrison.  Unfortunately, it seems that
it's not necessarily the case that every well-order has a canonical
equivalent.  For example, there may be no canonical version of ω+1
because ω may have already exhausted all the underlying carrier set.

If I'm right, this rather reduces the appeal of this subset.