compiler correctness progress: finish Cevaluate_any_env
i.e. that you can add and remove junk to environments in the
intermediate language semantics.

make doPrim2 return an option so it becomes totally specified

experimentation with THEN1 instead of TRY in the proofscript, and
abstracting out some common bits of tactic. the proof still probably
takes longer than it should to replay.