fix error in definition of free_vars on match expressions
(In the intermediate language.)
Previously neglected the fact that the variables are bound by the
pattern, so was counting them as free.

Also, strengthened the theorem about the environment resulting from a
successful pattern match: its domain is not just a subset of the pattern
variables (plus the incoming accumulator), but is exactly equal to the
pattern variables (plus accumulator). That is, every pattern variable
gets bound, whether or not they then appear free in the associated
expression.

Found this bug by trying to prove that if the intermediate language
semantics returns a value, then the free vars of the expression are all
in the domain of the environment.