compiler correctness attempts
Tried to prove that if Cevaluate produces a result then the free
variables of the expression were all in the environment, but that is not
true because of lazy evaluation (of things like pattern-matching and if
statements). Of course in reality the whole expression would have been
type-checked first...

Tried to make progress on exp_to_Cexp_thm1, but the statement may need
to change to either force the environment in the conclusion to have the
free variables of the expression, or to make use of the fact that
evaluate succeeded (i.e. no type error)...