get compile testing infrastructure working again
by omitting the equations defining v_to_Cv from the mutually recursive
equations for exp_to_Cexp. (eventually we could define a new version of
exp_to_Cexp that doesn't have those equations and prove it equivalent on
source expressions to the one about which the theorems are proved.)

also tweaked underscore rule, because RENAME_VARS_CONV primes variables
as necessary. but we don't want underscores to be primed. so just rename
them all to "usn" for some number n.

surprisingly, the new tests succeed, so the treatment of declaration
variables may be correct. but it's still too complicated.