major changes to miniml compiler intermediate lang and wrapper
mostly trying to simplify treatement of declarations.

added a dedicated expression to the intermediate language to sit at the
bottom of a declaration, and reference all the declaration variables.

exp_to_Cexp no longer modifies the state, and no longer has its own
state type but just uses the repl state. it doesn't need to modify the
state because pat_to_Cpat and repl_dec take care of updating the map
with new declaration variables. (any other variable mappings made during
descent of the expression can be thrown away because they're local.)
remove number_pat_vars and extend_rs; abstract out two cases in
exp_to_Cexp for reuse in repl_dec.

compiling the new declaration expression toggles a flag in the compiler
state, which compile_bindings reads on return to determine whether to
pop stack variables or not. the compiler state no longer explicitly
tracks whether it was called on a declaration or not.

none of this is tested. (this commit is just a checkpoint.)