more changes to compiler and fix tests
issues were with how compile_bindings, compilation of a CDecl, and
branching expressions interact. the method now is: compilation of CDecl
saves the environment and stack size for the next repl invocation in a
separate slot, and otherwise looks like an expression that returned one
argument onto the stack. therefore things like CIf don't get messed up
by it being on one of the branches. compile_bindings doesn't pop the new
bindings if the body ends in a CDecl, because we leave it up to CDecl to
rearrange the stack completely (this behavior is as before the major
changes).