attempt to prove free_vars_exp_to_Cexp
Define FV and pat_vars for source language exps and pats.

Also, define exp_to_Cexp_nice_ind as to use as the proof method: this
was tricky to get right (and may still be wrong).

Although this proof is almost finished, it is already too long, and I am
thinking it will be better to replace more (or all) occurrences of FOLDR
in the of exp_to_Cexp with explicitly named functions. The advantage I
predict is better induction theorems and more constants around which to
structure the theory.

This commit also includes a "cases"-like theorem for exp_to_Cexp that
will probably not be useful in the end. (It was used when originally
trying the proof by free_vars_ind.)