define recursive notion of properly closed MiniML closures
if the free variables of the body are contained in the argument +
environment domain, and also every other value (e.g., in the environment,
recursively) is also closed like that.

using the generic mk_EVERY code that was used to define canonical_env on
intermediate language values similarly. although it's not completely
generic: I have hard-coded how much recursive searching for appropriate
types it does (currently just lists and lists of pairs or triples with
recursion in the second half of the triple).