fix bug in remove_mat, simplify remove_mat
The problem was that the variable bound to the value being matched
(against which projections would be made to match pieces) could be
shadowed by a pattern variable. Now, we make sure to pick a variable
that is fresh for all body and pattern variables.

Also, don't reuse the variable if the match was already against a
variable (since that variable might not be fresh enough). This
simplifies things since we don't branch on whether it's a variable or
not.