Correct right-shift operation in Crypt/TEA.
Now use Logical shift right (in HOL >>>). Before, arithmetic shift
right (in HOL >>) was used. It looked like the C I was working from,
but had different semantics. Once again a reason to hate and distrust
overloading (in C this time, but its good to hate overloading
universally).

Thanks to David Hardin for noticing this bug.