Prove the basic polynomial representation theorem for ordinals.
Namely, given a base a that is bigger than 1, every ordinal can be
expressed as a finite sum that looks like

  c1 * a^e1 + c2 * a^e2 + c3 * a^e3 ..

where each exponent is bigger than all following exponents and each
coefficient is positive and less than a. For example, if a = 2, then
the representation of omega + 4 is

  2^omega + 2^2

Need to still prove that the polynomial representation is unique.