Fix PFset_conv to not quote terms in one of its internal proofs.
The problem with this is that later theories may overload "SIGMA" or
other constants, thereby changing the meaning of this code when it's
evaluated. (Indeed, this is exactly what code in src/probability did
do.) Good practice, thus, is not to use the parser in library code, or
to rebind the parser so that it is not using the global grammar.