work on pattern matching correctness
exp_to_Cexp_thm1 is almost ready now for application of as-yet-unproved
theorem relating Cevaluate_match and remove_mat_var. The latter's
statement probably needs to be changed to only reason up to syneq and to
state exactly which variables must be free for what or contained in
what before it can be proved.