Don't use fancy UTF8 quotes when converting text help files on Windows.
I did look into the possibility of changing the help system to
dynamically convert quotes when coming to display the file (if the
Unicode flag was off, you could convert the fancy quotes back to pure
ASCII as the file was displayed). But there doesn't seem to be a
good/easy way of getting into the internals of the Moscow ML's help
system's display routines for this functionality. And of course, it's
people using Moscow ML on Windows that are most likely to have trouble
displaying the fancy quotes. So, this is a somewhat hackish way out.
We can revisit if and when we drop support for Moscow ML.

Closes #40.