Cpnomatch_remove_mat attempt 3
made Cpnomatch_list require lists to be the same length, so it doesn't
overlap with Cpmatch_error (but haven't proved that).
also, make Cpnomatch_list test patterns in order, and ensure all
patterns up to the first non-matching one match. This is perhaps
overspecifying its behaviour, but works for now.