work on Cevaluate_match_remove_mat_var
change the statement slightly to avoid a case split in the second
subgoal, but the new statement requires Cevaluate to be deterministic.
In fact, Cevaluate is not currently deterministic because non-recursive
closures are allowed to use any fresh dummy name to name themselves;
this should be changed to pick a particular fresh name, since that
freedom doesn't buy much.
(This commit includes an attempt to prove Cevaluate_determ that failed.)