make Defn.add_persistent_funs play nice with new compLib version (#73)
compLib.add_persistent_funs previously accepted a string that was
arbitrary code to be printed at the end of a theory. Now it must be the
name of a theorem in the current theory.

Defn was running SUC_TO_NUMERAL_DEFN_CONV in the printed code; now it
saves a theorem with the result of the conversion and uses its name
instead.