Update dependencies in examples/separationLogic/src/Holmakefile
The issue is that if a Holmakefile specifies commands for building a
target (like vars_as_resourceBaseFunctor.uo in this case), then the
Holmakefile takes precedence for dependency analysis (i.e., no
automatic dependency analysis is done). In Poly/ML, calculated
dependencies are then the basis for the construction of the .uo file.
The .uo file specifies what gets loaded when load is used. In this
case, without the correct dependencies listed, an attempt to do load
"vars_as_resourceBaseFunctor" will fail with complaints about not
being able to resolve names of modules (listLib and bagLib in this
case).